strings: Refix 329ebdff: Enlarge data type
Make val in valstr 32-bit to accomodate the values added
by commit 329ebdff.

+++ include/ipmitool/helper.h
@@ -63,7 +63,7 @@
 struct ipmi_intf;
 
 struct valstr {
-	uint16_t val;
+	uint32_t val;
 	const char * str;
 };
 struct oemvalstr {
